backend/requirements.txt
absl-py==2.3.1
accelerate==0.0.1
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.3
altair==5.2.0
annotated-doc==0.0.3
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.11.0
APScheduler==3.10.4
astunparse==1.6.3
attrs==25.4.0
backoff==2.2.1
bandit==1.9.2
bcrypt==4.0.1
beautifulsoup4==4.14.3
bidict==0.23.1
bleach==6.3.0
blinker==1.9.0
boolean.py==5.0
boto3==1.42.17
botocore==1.42.17
bracex==2.6
brotli==1.2.0
CacheControl==0.14.4
cachetools==6.2.2
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.4.4
click==8.3.0
cohere==4.37
colorama==0.4.6
ConfigArgParse==1.7.1
contourpy==1.2.0
coverage==7.13.1
cryptography==46.0.0
cycler==0.12.1
cyclonedx-python-lib==11.6.0
deepface==0.0.96
defusedxml==0.7.1
Deprecated==1.3.1
distro==1.9.0
dnspython==2.7.0
dockerfile-parse==2.0.1
docstring_parser==0.17.0
e2b==2.9.0
e2b-code-interpreter==2.4.1
ecdsa==0.19.1
email-validator==2.1.0
et_xmlfile==2.0.0
fakeredis==2.20.1
fastapi==0.120.3
fastavro==1.11.1
ffmpy==0.3.1
filelock==3.20.1
fire==0.7.1
firebase_admin==7.1.0
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.7.0
fsspec==2023.12.2
gast==0.7.0
gdown==5.2.0
gevent==25.9.1
geventhttpclient==2.3.5
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.48.0
google-auth-httplib2==0.2.1
google-cloud-core==2.5.0
google-cloud-firestore==2.22.0
google-cloud-storage==3.7.0
google-crc32c==1.8.0
google-genai==1.62.0
google-generativeai==0.8.5
google-pasta==0.2.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
gradio==6.2.0
gradio_client==2.0.2
greenlet==3.2.3
groovy==0.1.2
groq==0.4.1
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
h5py==3.15.1
hpack==4.1.0
http_ece==1.2.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.34.0
hyperframe==6.1.0
idna==3.11
importlib-metadata==6.11.0
importlib-resources==6.1.1
iniconfig==2.3.0
intelhex==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.3
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jwcrypto==1.5.6
keras==3.13.0
kiwisolver==1.4.5
libclang==18.1.1
license-expression==30.4.4
lightecc==0.0.4
lightphe==0.0.20
limits==4.2
locust==2.42.2
locust-cloud==1.28.1
lxml==6.0.2
lz4==4.4.5
Mako==1.3.10
mando==0.7.1
Markdown==3.10
markdown-it-py==3.0.0
markdown2==2.5.4
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
msgpack==1.1.2
mtcnn==1.0.0
multidict==6.6.3
namex==0.1.0
networkx==3.6.1
nltk==3.9.2
numpy==1.26.2
openai==1.95.1
opencv-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.5
packageurl-python==0.17.6
packaging==24.1
pandas==2.1.4
passlib==1.7.4
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.0.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.10.0
platformdirs==4.5.0
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.3
psycopg2-binary==2.9.9
py-serializable==2.1.0
py-vapid==1.9.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.3
pydantic_core==2.41.4
pydub==0.25.1
Pygments==2.17.2
PyJWT==2.10.1
PyLTI1p3==2.0.0
PyMySQL==1.1.1
pyparsing==3.1.1
pypdfium2==5.3.0
PySocks==1.7.1
pytest==8.4.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
python-dateutil==2.8.2
python-docx==1.2.0
python-dotenv==1.2.1
python-engineio==4.12.2
python-http-client==3.3.7
python-jose==3.5.0
python-multipart==0.0.21
python-socketio==5.14.3
pytz==2023.3.post1
pywebpush==2.1.2
pywin32==311
PyYAML==6.0.1
pyzmq==27.1.0
radon==6.0.1
redis==7.0.1
referencing==0.32.0
regex==2025.11.3
reportlab==4.4.5
requests==2.32.5
retina-face==0.0.17
rich==14.0.0
rpds-py==0.15.2
rsa==4.9.1
ruff==0.14.8
s3transfer==0.16.0
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
semantic-version==2.10.0
sendgrid==6.12.5
sentry-sdk==2.48.0
setuptools==80.9.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.1
spotipy==2.25.1
SQLAlchemy==2.0.41
starlette==0.49.1
stevedore==5.6.0
sympy==1.14.0
tenacity==9.1.4
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
textblob==0.19.0
tf_keras==2.20.1
threadpoolctl==3.6.0
tokenizers==0.22.1
tomli==2.3.0
tomli_w==1.2.0
tomlkit==0.12.0
toolz==0.12.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
typer==0.15.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2023.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
vaderSentiment==3.3.2
wcmatch==10.1
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.0
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
zope.event==6.1
zope.interface==8.1
